+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/block/SpectreTreeBlocks.kt
@@ -60,33 +60,9 @@ object SpectreTreeBlocks {
 				.mapColor(MapColor.TERRACOTTA_LIGHT_BLUE)
 				.noOcclusion()
 		) {
-			override fun getToolModifiedState(state: BlockState, context: UseOnContext, itemAbility: ItemAbility, simulate: Boolean): BlockState? {
-				if (itemAbility != ItemAbilities.AXE_STRIP) return super.getToolModifiedState(state, context, itemAbility, simulate)
-
-				return ModBlocks.STRIPPED_SPECTRE_LOG.get()
-					.defaultBlockState()
-					.setValue(AXIS, state.getValue(AXIS))
-			}
-
 			override fun skipRendering(state: BlockState, adjacentBlockState: BlockState, side: Direction): Boolean {
 				return adjacentBlockState.`is`(this)
 			}
-
-		}
-
-	val STRIPPED_SPECTRE_LOG: FlammableRotatedPillarBlock =
-		object : FlammableRotatedPillarBlock(
-			Blocks.STRIPPED_OAK_LOG,
-			Properties
-				.ofFullCopy(Blocks.STRIPPED_OAK_LOG)
-				.mapColor(MapColor.TERRACOTTA_LIGHT_BLUE)
-				.noOcclusion()
-		) {
-
-			override fun skipRendering(state: BlockState, adjacentBlockState: BlockState, side: Direction): Boolean {
-				return adjacentBlockState.`is`(this)
-			}
-
 		}
 
 	val SPECTRE_LEAVES: LeavesBlock =