+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/datagen/tag/ModBlockTagsProvider.kt
@@ -49,15 +49,9 @@ class ModBlockTagsProvider(
 				ModBlocks.SPECTRE_LEAVES.get()
 			)
 
-		this.tag(C_STRIPPED_WOODS)
-			.add(
-				ModBlocks.STRIPPED_SPECTRE_LOG.get()
-			)
-
 		this.tag(BlockTags.LOGS_THAT_BURN)
 			.add(
 				ModBlocks.SPECTRE_LOG.get(),
-				ModBlocks.STRIPPED_SPECTRE_LOG.get(),
 				ModBlocks.SPECTRE_WOOD.get()
 			)
 
@@ -402,7 +396,6 @@ class ModBlockTagsProvider(
 				ModBlocks.CHERRY_PLATFORM.get(),
 				ModBlocks.SPECTRE_PLANKS.get(),
 				ModBlocks.SPECTRE_LOG.get(),
-				ModBlocks.STRIPPED_SPECTRE_LOG.get(),
 				ModBlocks.BEAN_STALK.get(),
 				ModBlocks.LESSER_BEAN_STALK.get(),
 				ModBlocks.CUSTOM_CRAFTING_TABLE.get(),