+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 - Fixed a mysterious NPE that sometimes happens with the Block Breaker maybe (#59)
 - Fixed Spectre Cubes thinking they're somewhere else, so second+ players wouldn't be able to use them (#57)
 
+### Removed
+
+- Stripped Spectre Logs
+	- I don't want to make a texture
+
 # 1.8.1
 
 ### Fixed